@@ -24,13 +24,13 @@
     "moment": "2.22.2",
     "rc-slider": "8.6.3",
     "rc-time-picker": "3.6.2",
-    "react": "16.7.0",
+    "react": "16.8.4",
     "react-autosuggest": "9.4.3",
     "react-beautiful-dnd": "10.0.3",
     "react-copy-to-clipboard": "5.0.1",
     "react-css-modules": "4.7.9",
     "react-custom-scrollbars": "4.2.1",
-    "react-dom": "16.8.1",
+    "react-dom": "16.8.4",
     "react-flip-move": "3.0.3",
     "react-icons": "3.2.2",
     "react-jsx-highcharts": "3.3.0",
